+++ hc/api/tests/test_notify_rocketchat.py
@@ -65,14 +65,13 @@ class NotifyRocketChatTestCase(BaseTestCase):
     def test_it_shows_schedule_and_tz(self, mock_post):
         mock_post.return_value.status_code = 200
         self.check.kind = "cron"
-        self.check.schedule = "1 1 1 1 1"
         self.check.tz = "Europe/Riga"
         self.check.save()
 
         self.channel.notify(self.check)
         attachment = mock_post.call_args.kwargs["json"]["attachments"][0]
         fields = {f["title"]: f["value"] for f in attachment["fields"]}
-        self.assertEqual(fields["Schedule"], "1 1 1 1 1")
+        self.assertEqual(fields["Schedule"], "\u034f* \u034f* \u034f* \u034f* \u034f*")
         self.assertEqual(fields["Time Zone"], "Europe/Riga")
 
     @patch("hc.api.transports.curl.request")