+++ hc/api/tests/test_notify_slack.py
@@ -93,7 +93,6 @@ class NotifySlackTestCase(BaseTestCase):
     def test_it_shows_schedule_and_tz(self, mock_post):
         self._setup_data("123")
         self.check.kind = "cron"
-        self.check.schedule = "1 1 1 1 1"
         self.check.tz = "Europe/Riga"
         self.check.save()
         mock_post.return_value.status_code = 200
@@ -102,7 +101,7 @@ class NotifySlackTestCase(BaseTestCase):
         payload = mock_post.call_args.kwargs["json"]
         attachment = payload["attachments"][0]
         fields = {f["title"]: f["value"] for f in attachment["fields"]}
-        self.assertEqual(fields["Schedule"], "1 1 1 1 1")
+        self.assertEqual(fields["Schedule"], "\u034f* \u034f* \u034f* \u034f* \u034f*")
         self.assertEqual(fields["Time Zone"], "Europe/Riga")
 
     @patch("hc.api.transports.curl.request")