+++ hc/front/forms.py
@@ -336,6 +336,7 @@ class NtfyForm(forms.Form):
     error_css_class = "has-error"
     topic = forms.CharField(max_length=50)
     url = LaxURLField(max_length=1000)
+    token = forms.CharField(max_length=100, required=False)
     priority = forms.IntegerField(initial=3, min_value=0, max_value=5)
     priority_up = forms.IntegerField(initial=3, min_value=0, max_value=5)