+++ hc/api/tests/test_notify_ntfy.py
@@ -8,7 +8,7 @@ from unittest.mock import patch
 
 from django.utils.timezone import now
 
-from hc.api.models import Channel, Check, Notification
+from hc.api.models import Channel, Check, Notification, Ping
 from hc.test import BaseTestCase
 
 
@@ -18,9 +18,16 @@ class NotifyNtfyTestCase(BaseTestCase):
 
         self.check = Check(project=self.project)
         self.check.name = "Foo"
+        self.check.tags = "foo bar"
+        self.check.n_pings = 123
         self.check.status = "down"
         self.check.last_ping = now() - td(minutes=61)
         self.check.save()
+
+        self.ping = Ping(owner=self.check)
+        self.ping.n = 1
+        self.ping.remote_addr = "1.2.3.4"
+        self.ping.save()
 
         self.channel = Channel(project=self.project)
         self.channel.kind = "ntfy"
@@ -44,8 +51,27 @@ class NotifyNtfyTestCase(BaseTestCase):
 
         payload = mock_post.call_args.kwargs["json"]
         self.assertEqual(payload["title"], "Foo is DOWN")
+        self.assertIn("Project: Alices Project", payload["message"])
+        self.assertIn("Tags: foo, bar", payload["message"])
+        self.assertIn("Period: 1 day", payload["message"])
+        self.assertIn("Total Pings: 123", payload["message"])
+        self.assertIn("Last Ping: Success, now", payload["message"])
+
         self.assertEqual(payload["actions"][0]["url"], self.check.cloaked_url())
         self.assertNotIn("All the other checks are up.", payload["message"])
+
+    @patch("hc.api.transports.curl.request")
+    def test_it_shows_schedule_and_tz(self, mock_post):
+        mock_post.return_value.status_code = 200
+
+        self.check.kind = "cron"
+        self.check.tz = "Europe/Riga"
+        self.check.save()
+        self.channel.notify(self.check)
+
+        payload = mock_post.call_args.kwargs["json"]
+        self.assertIn("Schedule: * * * * *", payload["message"])
+        self.assertIn("Time Zone: Europe/Riga", payload["message"])
 
     @patch("hc.api.transports.curl.request")
     def test_it_shows_all_other_checks_up_note(self, mock_post):