+++ hc/api/transports.py
@@ -1190,7 +1190,11 @@ class Ntfy(HttpTransport):
         return self.priority(check) == 0
 
     def notify(self, check, notification=None) -> None:
-        ctx = {"check": check, "down_checks": self.down_checks(check)}
+        ctx = {
+            "check": check,
+            "ping": self.last_ping(check),
+            "down_checks": self.down_checks(check),
+        }
         payload = {
             "topic": self.channel.ntfy_topic,
             "priority": self.priority(check),