+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Improvements
 - Add Channel.last_notify_duration field, use it in "sendalerts" for prioritization
 - Update Telegram integration to treat "bot was blocked by the user" as permanent error
+- Add "Time Zone" field in notifications that use the "Schedule" field (#863)
 
 ### Bug Fixes
 - Fix "senddeletionnotices" to recognize "Supporter" subscriptions